[WIP] Domains: Add availability precheck before adding a domain suggestion to the cart
Since we may get some false positives from the domain suggestions engines we should better handle domains that are actually not available. So the idea is to add a pre-check for the domain and not add it to the cart if it's already registered.

I still have to handle subdomains and some other cases